Allow non-admin users to execute SHOW DATABASES
This commit introduces a new interface type, influxql.ResourceShower,
that is used to determine whether a resource can be returned in a SHOW
query (it currently only supports SHOW DATABASES). When authentication
is enabled, admin users can see all databases, and non-admin users can
see only databases for which they have read and/or write permission.
When authentication is disabled, all databases are visible.

This addresses a long-standing issue where Chronograf or Grafana would
be unable to list databases if the logged-in user did not have admin
privileges.

Fixes #4785.